backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
coverage==4.5.4
cryptography==44.0.3
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-modeltranslation==0.19.14
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
firebase-admin==6.8.0
future==1.0.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
serializers==0.2.4
service-identity==24.2.0
setuptools==80.3.1
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zope.interface==7.2
